scripts/requirements.txt
duckdb==1.0.0
google_api_python_client==2.136.0
langchain==0.2.6
langchain_community==0.2.6
langchain_core==0.2.10
openai_whisper==20231117
pandas==2.2.2
psycopg2-binary
yt_dlp==2024.7.1
